makrcave-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
qrcode[pil]==7.4.2
Pillow==10.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
alembic==1.12.1
psycopg2-binary==2.9.9
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
email-validator==2.1.0
reportlab==4.0.7
qrcode[pil]==7.4.2
razorpay==1.3.0
stripe==7.8.0
bleach==6.1.0
